[Vue3 + TS + Quasar] refactoring: App and NavHeader now use the component aggregator for component imports
